% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copresence.R
\name{co.p}
\alias{co.p}
\title{Co-Presence - from Peeples 2017, http://mattpeeples.net/netstats.html. Creates a square matrix with 1 and 0 for presence and absence.}
\usage{
co.p(x, thresh = 0.1)
}
\arguments{
\item{x}{A dataset of count variables}

\item{thresh}{Threshold. Anything above this value will be counted as present.}
}
\value{
A matrix
}
\description{
Co-Presence - from Peeples 2017, http://mattpeeples.net/netstats.html. Creates a square matrix with 1 and 0 for presence and absence.
}
\examples{
co.p(df, thresh = 20)
}
