% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeedgelist.R
\name{make.edgelist}
\alias{make.edgelist}
\title{Make an Edgelist.his function makes an edgelist out of an archaeological dataset by using the row names as nodes, and the rest of the columns as counts of types.}
\usage{
make.edgelist(data)
}
\arguments{
\item{data}{Dataset with nodes as row names}
}
\value{
A tibble
}
\description{
Make an Edgelist.his function makes an edgelist out of an archaeological dataset by using the row names as nodes, and the rest of the columns as counts of types.
}
\examples{

# Make a dataset with  row names (nodes) as a separate column
}
