% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mbk3edges}
\alias{mbk3edges}
\title{Michelsberg Phase 3 ceramics edgelist}
\format{
A data frame with 268 rows and 3 columns.
\describe{
  \item{from}{site name}
  \item{to}{site name}
  \item{weight}{number of shared ceramic types}
}
}
\source{
\url{https://cran.r-project.org/web/packages/archdata/index.html}
}
\usage{
mbk3edges
}
\description{
A dataframe representing a weighted edge list of ceramic co-occurrence in Michelsberg Phase 3.
}
\keyword{datasets}
