% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mbk3mat}
\alias{mbk3mat}
\title{Michelsberg Phase 3 ceramics matrix}
\format{
A 24x24 symmetric matrix.
}
\source{
\url{https://cran.r-project.org/web/packages/archdata/index.html}
}
\usage{
mbk3mat
}
\description{
A matrix showing the amount of ceramic types co-occurring between sites of Michelsberg Phase 3.
}
\keyword{datasets}
