% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mbk3nodes}
\alias{mbk3nodes}
\title{Michelsberg Phase 3 ceramics node list}
\format{
A data frame with 24 rows and 5 columns.
\describe{
  \item{id}{site name and feature number}
  \item{site_name}{site name}
  \item{feature_nr}{feature number}
  \item{lon}{longitude}
  \item{lat}{latitude}
}
}
\source{
\url{https://cran.r-project.org/web/packages/archdata/index.html}
}
\usage{
mbk3nodes
}
\description{
A dataframe representing a node list of ceramic find contexts in Michelsberg Phase 3.
}
\keyword{datasets}
